move MapKeySerializer implementation
